lxc monitor for RD

+++ main.py
@@ -2,5 +2,6 @@ from client.main import main
 import os
 
 path= os.path.join(os.path.dirname(__file__), 'log_shelf.db')
-os.remove(path)
+if(os.path.exists(path)):
+    os.remove(path)
 main.main()
